+++ libs/components/src/lib/components/dropdowns/multi-select/input-multi-select.component.ts
@@ -19,13 +19,13 @@ import {
   PrizmAutoFocusModule,
   PrizmDropdownControllerModule,
   PrizmFocusableModule,
-  PrizmHintDirective,
+  PrizmHintOnOverflowDirective,
   PrizmLifecycleDirective,
 } from '../../../directives';
 import { PRIZM_MULTI_SELECT_OPTIONS, PrizmMultiSelectOptions } from './multi-select.options';
 import { PrizmContextWithImplicit, PrizmNativeFocusableElement } from '../../../types';
 import { PrizmInputControl, PrizmInputNgControl, PrizmInputTextModule } from '../../input';
-import { prizmIsNativeFocused, prizmIsTextOverflow$ } from '../../../util';
+import { prizmIsNativeFocused } from '../../../util';
 import { debounceTime, map, shareReplay, startWith, switchMap, takeUntil, tap } from 'rxjs/operators';
 import { BehaviorSubject, combineLatest, Observable, Subject, timer } from 'rxjs';
 import { prizmDefaultProp } from '@prizm-ui/core';
@@ -70,7 +70,6 @@ import { PrizmIconsFullRegistry } from '@prizm-ui/icons/core';
     ReactiveFormsModule,
     CommonModule,
     PrizmLetDirective,
-    PrizmHintDirective,
     PrizmCallFuncPipe,
     PrizmAutoFocusModule,
     PrizmScrollbarComponent,
@@ -81,6 +80,7 @@ import { PrizmIconsFullRegistry } from '@prizm-ui/icons/core';
     PrizmDropdownHostComponent,
     PrizmIconsFullComponent,
     PrizmFocusableModule,
+    PrizmHintOnOverflowDirective,
   ],
   providers: [
     {
@@ -203,7 +203,6 @@ export class PrizmInputMultiSelectComponent<T> extends PrizmInputNgControl<T[]> 
   }
 
   public readonly defaultIcon = 'triangle-down';
-  readonly prizmIsTextOverflow$ = prizmIsTextOverflow$;
   public readonly direction: PrizmOverlayOutsidePlacement = PrizmOverlayOutsidePlacement.RIGHT;
 
   public readonly items$ = new BehaviorSubject([]);